requirements.txt
beautifulsoup4 
playwright 
spacy 
requests 
pydantic
python-dotenv
playwright-stealth